Integrate color-change regression tests into main suite
